refactor: data 모듈 유틸 클래스 경로 이동
