+++ android/Bottari/data/src/main/java/com/bottari/data/common/extension/ResponseExtension.kt
@@ -1,4 +1,4 @@
-package com.bottari.data.extension
+package com.bottari.data.common.extension
 
 import retrofit2.Response
