+++ android/Bottari/data/src/main/java/com/bottari/data/common/util/LocalDateSerializer.kt
@@ -1,20 +1,20 @@
-package com.bottari.data.util
+package com.bottari.data.common.util
 
 import kotlinx.serialization.KSerializer
 import kotlinx.serialization.descriptors.PrimitiveKind
 import kotlinx.serialization.descriptors.PrimitiveSerialDescriptor
 import kotlinx.serialization.encoding.Decoder
 import kotlinx.serialization.encoding.Encoder
-import java.time.LocalTime
+import java.time.LocalDate
 
-object LocalTimeSerializer : KSerializer<LocalTime> {
-    private const val SERIAL_NAME = "LocalTime"
+object LocalDateSerializer : KSerializer<LocalDate> {
+    private const val SERIAL_NAME = "LocalDate"
     override val descriptor = PrimitiveSerialDescriptor(SERIAL_NAME, PrimitiveKind.STRING)
 
     override fun serialize(
         encoder: Encoder,
-        value: LocalTime,
+        value: LocalDate,
     ) = encoder.encodeString(value.toString())
 
-    override fun deserialize(decoder: Decoder): LocalTime = LocalTime.parse(decoder.decodeString())
+    override fun deserialize(decoder: Decoder): LocalDate = LocalDate.parse(decoder.decodeString())
 }
