+++ android/Bottari/data/src/main/java/com/bottari/data/common/util/PrettyJsonLogger.kt
@@ -1,4 +1,4 @@
-package com.bottari.data.util
+package com.bottari.data.common.util
 
 import kotlinx.serialization.ExperimentalSerializationApi
 import kotlinx.serialization.json.Json
@@ -46,7 +46,11 @@ class PrettyJsonLogger : HttpLoggingInterceptor.Logger {
         (text.startsWith(JSON_OBJECT_START) && text.endsWith(JSON_OBJECT_END)) ||
             (text.startsWith(JSON_ARRAY_START) && text.endsWith(JSON_ARRAY_END))
 
-    private fun isLastLine(text: String): Boolean = text.startsWith(REQUEST_END_PREFIX) || text.startsWith(RESPONSE_END_PREFIX)
+    private fun isLastLine(text: String): Boolean =
+        text.startsWith(REQUEST_END_PREFIX) ||
+            text.startsWith(
+                RESPONSE_END_PREFIX,
+            )
 
     companion object {
         private const val TAG = "NETWORK"
