+++ android/Bottari/data/src/main/java/com/bottari/data/common/util/SafeApiCall.kt
@@ -1,4 +1,4 @@
-package com.bottari.data.util
+package com.bottari.data.common.util
 
 import retrofit2.HttpException
 import retrofit2.Response
