+++ android/Bottari/data/src/main/java/com/bottari/data/source/remote/BottariItemRemoteDataSourceImpl.kt
@@ -1,9 +1,9 @@
 package com.bottari.data.source.remote
 
+import com.bottari.data.common.util.safeApiCall
 import com.bottari.data.model.item.FetchChecklistResponse
 import com.bottari.data.model.item.SaveBottariItemsRequest
 import com.bottari.data.service.BottariItemService
-import com.bottari.data.util.safeApiCall
 
 class BottariItemRemoteDataSourceImpl(
     private val bottariItemService: BottariItemService,
