+++ android/Bottari/data/src/main/java/com/bottari/data/source/remote/BottariTemplateRemoteDataSourceImpl.kt
@@ -1,11 +1,11 @@
 package com.bottari.data.source.remote
 
-import com.bottari.data.extension.extractIdFromHeader
+import com.bottari.data.common.extension.extractIdFromHeader
+import com.bottari.data.common.util.safeApiCall
 import com.bottari.data.model.template.CreateBottariTemplateRequest
 import com.bottari.data.model.template.FetchBottariTemplateResponse
 import com.bottari.data.model.template.FetchMyBottariTemplatesResponse
 import com.bottari.data.service.BottariTemplateService
-import com.bottari.data.util.safeApiCall
 
 class BottariTemplateRemoteDataSourceImpl(
     private val bottariTemplateService: BottariTemplateService,
